Make download headers customizable
Before, CSV downloads would be sent with a mime type of 'text/plain'.
This made Firefox propose to open the file with a text editor instead of LibreOffice / Excel.

The workaround without this fix would be a bit longwinded:

```php
	$response = Excel::download(
            new MyExport($data), "{$filename}.csv",
            \Maatwebsite\Excel\Excel::CSV
        );

	// send the correct mime type
        $response->headers->set('Content-Type', 'text/csv');

        return $response;
```

With this fix:

```php
	return Excel::download(
	    new MyExport($data), "{$filename}.csv",
            \Maatwebsite\Excel\Excel::CSV,
	    ['Content-Type' => 'text/csv']
        );
```

Doesn't break the API. Doesn't change behavior.  Custom headers are optional.