AfterImport change behaviour of getDelegate.
Based on the Documentation:

"Both Reader and Sheet have a ->getDelegate() method which returns the underlying PhpSpreadsheet class. This will allow you to add custom macros as shortcuts to PhpSpreadsheet methods that are not available in this package."

Please confirm if It applies to other Events. _(Same change proposed for AfterImport at the moment)_

Reader.php returns 
- a Workbook Object when calling getDelegate()
- a IReader Object when calling getPhpSpreadsheetReader()

Result: 

```
Xlsx {#280 ▼
  -referenceHelper: ReferenceHelper {#279}
  #readDataOnly: true
  #readEmptyCells: true
  #includeCharts: false
  #loadSheetsOnly: array:2 [▶]
  #readFilter: DefaultReadFilter {#271}
  #fileHandle: null
  #securityScanner: XmlScanner {#291 ▶}
}

```